Trying to merge Monday

+++ monolith_service/routers/produce.py
@@ -47,16 +47,16 @@ def update_produce(
     produce_id: int,
     produce: Produce_create,
     queries: ProduceQueries = Depends(),
-    # account_data: dict = Depends(authenticator.get_current_account_data)
+    account_data: dict = Depends(authenticator.get_current_account_data)
 ) -> Union[Produce_get, Error]:
-    # if "admin" in account_data.get("username"):
+    if "admin" in account_data.get("username"):
         return queries.update_produce(produce_id, produce)
-    # else:
-    #     raise HTTPException(
-    #                 status_code=status.HTTP_401_UNAUTHORIZED,
-    #                 detail="Invalid token",
-    #                 headers={"WWW-Authenticate": "Bearer"},
-                # )
+    else:
+        raise HTTPException(
+                    status_code=status.HTTP_401_UNAUTHORIZED,
+                    detail="Invalid token",
+                    headers={"WWW-Authenticate": "Bearer"},
+                )
 
 @router.delete("/api/produce/{produce_id}/delete", response_model = bool)
 def delete_produce(
@@ -65,15 +65,15 @@ def delete_produce(
     account_data: dict = Depends(authenticator.get_current_account_data)
 )-> bool:
     print(account_data)
-    # if "admin" in account_data.get("username"):
-    return queries.delete_produce(produce_id)
-    # else:
-    #     print(account_data)
-    #     raise HTTPException(
-    #                 status_code=status.HTTP_401_UNAUTHORIZED,
-    #                 detail="Invalid token",
-    #                 headers={"WWW-Authenticate": "Bearer"},
-                # )
+    if "admin" in account_data.get("username"):
+        return queries.delete_produce(produce_id)
+    else:
+        print(account_data)
+        raise HTTPException(
+                    status_code=status.HTTP_401_UNAUTHORIZED,
+                    detail="Invalid token",
+                    headers={"WWW-Authenticate": "Bearer"},
+                )
 
 
 @router.patch("/api/produce/{produce_id}/patch", response_model = Produce_get)
@@ -81,23 +81,23 @@ def update_produce_available(
     produce_id: int,
     produce: Produce_update_available,
     queries: ProduceQueries = Depends(),
-    # account_data: dict = Depends(authenticator.get_current_account_data)
+    account_data: dict = Depends(authenticator.get_current_account_data)
 ) -> Produce_get:
     print(produce)
-    # if "admin" in account_data.get("username"):
-    return queries.update_produce_available(produce_id, Produce_update_available(
-    product_name = produce.product_name,
-    picture_file = produce.picture_file,
-    available = produce.available,
-    height = produce.height, 
-    length = produce.length, 
-    width = produce.width
-    ))
-    # else:
-    #     raise HTTPException(
-    #                 status_code=status.HTTP_401_UNAUTHORIZED,
-    #                 detail="Invalid token",
-    #                 headers={"WWW-Authenticate": "Bearer"},
-    #             )
+    if "admin" in account_data.get("username"):
+        return queries.update_produce_available(produce_id, Produce_update_available(
+        product_name = produce.product_name,
+        picture_file = produce.picture_file,
+        available = produce.available,
+        height = produce.height, 
+        length = produce.length, 
+        width = produce.width
+        ))
+    else:
+        raise HTTPException(
+                    status_code=status.HTTP_401_UNAUTHORIZED,
+                    detail="Invalid token",
+                    headers={"WWW-Authenticate": "Bearer"},
+                )
